Corrected parser und factories

+++ ImplementationPhase/Model/ParsingPackage/include/Factory/ConvolutionalLayerFactory.hpp
@@ -2,7 +2,6 @@
 #define CONVOLUTIONAL_LAYER_FACTORY_H_
 
 #include "LayerFactory.hpp"
-#include "ConvolutionalLayer.hpp"
 #include "NeuralNetworkLayer.hpp"
 
 #include <vector>
@@ -11,12 +10,12 @@ class ConvolutionalLayerFactory : public LayerFactory
 {
 public:
     NeuralNetworkLayer buildLayer() override;
-    ConvolutionalLayerFactory setWeightTensors(std::vector<std::vector<std::vector<int>>> weightTensors);
+    ConvolutionalLayerFactory setWeightTensors(std::vector<std::vector<std::vector<float>>> weightTensors);
     PoolingLayConvolutionalLayerFactoryerFactory setStride(int stride);
     ConvolutionalLayerFactory setPadding(int padding);
 
 private:
-    std::vector<std::vector<std::vector<int>>> weightTensors;
+    std::vector<std::vector<std::vector<float>>> weightTensors;
     int stride;
     int padding;
 };
